♻️ Splitt opp VilkårperiodeContent i mindre komponenter i en egen vilkårperiodemappe
