+++ src/internt-vedtak/felles.tsx
@@ -15,14 +15,3 @@ export const Begrunnelse: React.FC<{ begrunnelse?: string }> = ({ begrunnelse })
         )
     );
 };
-
-export const KommentarSlettet: React.FC<{ data: { slettetKommentar?: string } }> = ({ data }) => {
-    return (
-        data.slettetKommentar && (
-            <p style={{ whiteSpace: 'pre-wrap' }}>
-                Kommentar slettet: <br />
-                {data.slettetKommentar}
-            </p>
-        )
-    );
-};
